+++ packages/client-auth/RELEASES.md
@@ -59,3 +59,13 @@ Based on:
 - [typescript v1.0.0-alpha.5] packages/client-auth
 ### Releases
 - [NPM v1.0.0-alpha.5] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.5 - packages/client-auth
+
+## 2024-07-16 00:09:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.335.2 (2.372.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.6] packages/client-auth
+### Releases
+- [NPM v1.0.0-alpha.6] https://www.npmjs.com/package/@formance/sdk-auth/v/1.0.0-alpha.6 - packages/client-auth